Adding necessary imports for log-in functionality
UseState: We will eventually need to set a loggedIn state to show access to the rest of the application.

UseNavigate: Upon login, user will be directed to the patients page.

UseForm: Added async form onsubmit.

UseCookie: User's session will be stored in cookie for logged state.

This project is a lot of work but I'm having fun =) Coding rules!