Reset migrations
During the move from Sqlite to MSSQL, I had to temporarily disable IDENTITY on the ID columns for all assessments since they were being set in the submission handler. As part of getting patient history working I re-enabled this, but the columns needed to be dropped and recreated for migrations to work which EF wouldn't do automatically. Since we're not actually running in prod with data to worry about saving at the moment, the faster choice was to drop migrations and create a new starting point with MSSQL and Patient History.